requirements.txt
pdf2image==1.16.3
pytesseract==0.3.10
pymupdf==1.23.8
transformers==4.38.2
sentence-transformers==2.5.1
spacy==3.7.4
scikit-learn==1.4.0
python-dateutil==2.9.0
python-Levenshtein==0.23.0

